% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_likelihood_weights.R
\name{calc_likelihood_weights}
\alias{calc_likelihood_weights}
\title{Calculate likelihood weights.}
\usage{
calc_likelihood_weights(years, weight.rate)
}
\arguments{
\item{years}{A vector of the years that weights must be calculated for.}

\item{weight.rate}{The rate used to calculate the weights. Each year y is assigned weight proportional to (1 - weight.rate) ^ (- year).}
}
\value{
A vector of weights.
}
\description{
Calculate likelihood weights.
}
\examples{
calc_likelihood_weights(1980:2010, .05)
}

