% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_maple_data_format.R
\name{check_maple_data_format}
\alias{check_maple_data_format}
\title{Check arguments are in the desired format: 18xN matrices with column names corresponding to years}
\usage{
check_maple_data_format(...)
}
\description{
Check arguments are in the desired format: 18xN matrices with column names corresponding to years
}

