% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fetch_cities_file}
\alias{fetch_cities_file}
\title{Get csv of cities to fill out}
\usage{
fetch_cities_file(file = "cities.csv", overwrite = FALSE)
}
\arguments{
\item{file}{Path to save file to}

\item{overwrite}{Overwrite file if it already exists}
}
\value{

}
\description{
Get csv of cities to fill out
}
