% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fetch_states_file}
\alias{fetch_states_file}
\title{Get csv of states to fill out}
\usage{
fetch_states_file(file = "states.csv", overwrite = FALSE)
}
\arguments{
\item{file}{Path to save file to}

\item{overwrite}{Overwrite file if it already exists}
}
\value{

}
\description{
Get csv of states to fill out
}
