% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_rates_matrix.R
\name{fitted_rates_matrix}
\alias{fitted_rates_matrix}
\title{A convenience function to extract the mortality matrix from a model fit.}
\usage{
fitted_rates_matrix(model.fit)
}
\arguments{
\item{model.fit}{A model fitted by INLA or Lee-Carter methods.}
}
\value{
A matrix containing death rate model fits, with rows corresponding to age groups and columns to years.
}
\description{
A convenience function to extract the mortality matrix from a model fit.
}
\examples{
x <- maple_fit_model(maple_models()[[20]], deaths = maple.deaths, population = maple.population, forecast.horizon = 5)
fitted_rates_matrix(x)
}

