% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maple_scores.R
\name{get_maple_scores}
\alias{get_maple_scores}
\title{Get posterior probability scores}
\usage{
get_maple_scores(fit)
}
\arguments{
\item{fit}{A list returned by fit_maple()}
}
\value{
A fit object (list)
}
\description{
This function allows you to compute posterior uncertainty and continuous phenotype scores
}
\keyword{Bayesian}
\keyword{spatial}
\keyword{transcriptomics}
