% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init}
\alias{init}
\title{Create necessary files to generate map}
\usage{
init(dir, overwrite = FALSE)
}
\arguments{
\item{dir}{Directory to save files to}
}
\value{

}
\description{
Create necessary files to generate map
}
