% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_fitted_matrix.R
\name{inla_fitted_matrix}
\alias{inla_fitted_matrix}
\title{A convenience function to extract the mortality matrix from a model fit.}
\usage{
inla_fitted_matrix(model.fit, variable)
}
\arguments{
\item{model.fit}{A model fitted by INLA.}

\item{variable}{Variable to be extracted from the fitted values matrix; mean, median, etc.}
}
\description{
A convenience function to extract the mortality matrix from a model fit.
}

