% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maple.deaths}
\alias{maple.deaths}
\title{A matrix containing sample deaths data for running the functions and examples in the maple package.}
\format{A 18 x 54 matrix
\describe{
  \item{rows}{5-year age groups, 0-4, 5-9, ..., 80-84, 85+}
  \item{columns}{Year, 1960-2013}
}}
\usage{
maple.deaths
}
\description{
A matrix containing sample deaths data for running the functions and examples in the maple package.
}
\keyword{datasets}

