% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAPLE_CPP}
\alias{MAPLE_CPP}
\title{MAPLE}
\usage{
MAPLE_CPP(
  zscorexin,
  zscoreyin,
  Sigmain,
  samplen1,
  samplen2,
  Gibbsnumberin,
  burninproportion,
  initial_betain,
  pi_beta_shape_in,
  pi_beta_scale_in,
  pi_c_shape_in,
  pi_c_scale_in,
  pi_1_shape_in,
  pi_1_scale_in,
  pi_0_shape_in,
  pi_0_scale_in,
  sigma2x_in,
  sigma2y_in,
  rho_in
)
}
\description{
MAPLE
}
