% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_fit_ensemble.R
\name{maple_fit_ensemble}
\alias{maple_fit_ensemble}
\title{Fit an ensemble of models}
\usage{
maple_fit_ensemble(deaths, population, forecast.horizon,
  models = maple_models(), num.draws = 1000, ax = NULL,
  num.threads = inla.getOption("num.threads"), verbose = TRUE)
}
\description{
Fit an ensemble of models
}

