% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_model_weights.R
\name{maple_model_weights}
\alias{maple_model_weights}
\title{Calculate model weights.}
\usage{
maple_model_weights(projection.errors)
}
\arguments{
\item{projection.errors}{A numeric vector of projection errors of individual models.}
}
\value{
A vector of model weights.
}
\description{
Calculate model weights.
}
\examples{
maple_model_weights(c(0.1, 1, 10))
maple_model_weights(c(5, 5))
}

