% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_models.R
\name{maple_models}
\alias{maple_models}
\title{Generate a list of forecasting models.}
\usage{
maple_models(fixed.prec = 0.001, gamma.shape = 1, gamma.rate = 0.001)
}
\arguments{
\item{fixed.prec}{The precision on the fixed effects (common intercept and slope) normal priors.}

\item{gamma.shape}{The shape parameter for the loggamma prior on the random effects. See ?INLA::inla.models for more details.}

\item{gamma.rate}{The rate (inverse scale) parameter for the loggamma prior on the random effects. See ?INLA::inla.models for more details.}
}
\value{
A list of forecasting models. Each model is a list with the following entries
\describe{
  \item{name}{A short name to identify the model.}
  \item{desc}{A description of the model.}
  \item{fml}{(only for models fitted in INLA) The formula passed to INLA when running the model.}
  \item{likelihood.weight.rate}{(only for weighted likelihood models) The rate used when calculating the likelihood weights. See ?calc_likelihood_weights for more details.}
  \item{num.pcs}{(only for Lee-Carter models) The number of principal components.}
}
}
\description{
Generate a list of forecasting models.
}
\examples{
maple_models()[c(1, 20)]
maple_models(fixed.prec = 1e-5, gamma.shape = .1, gamma.rate = .1)
}

