% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_projection_error.R
\name{maple_projection_error}
\alias{maple_projection_error}
\title{Calculate projection error}
\usage{
maple_projection_error(deaths, population, ax = NULL, fitted.rates)
}
\description{
Calculate projection error
}

