% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_sample_summaries.R
\name{maple_sample_summaries}
\alias{maple_sample_summaries}
\title{Calculate summaries for death rates, life expectancy and probability of dying.}
\usage{
maple_sample_summaries(samples)
}
\arguments{
\item{samples}{Posterior sample draws produced with maple_sample().}
}
\value{
A data frame containing the mean, median, sd, 2.5th and 97.5th percentiles of age-specific death rate, life expectancy and probability of dying in each year.
}
\description{
Calculate summaries for death rates, life expectancy and probability of dying.
}

