% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapme}
\alias{mapme}
\title{Generate a map of places you have been/lived}
\usage{
mapme(
  state_file,
  city_file,
  border_color = "#F2F1F1",
  city_color = "#EA3223",
  status_fill = c("#E5E5E5", "#afb6d2", "#A0BEA5")
)
}
\arguments{
\item{state_file}{Path to state file (see `fetch_states_file`)}

\item{city_file}{Path to city file (see `fetch_cities_file`)}

\item{border_color}{Border color}

\item{city_color}{City color}

\item{status_fill}{State color (length 2: never, visited, lived)}
}
\value{

}
\description{
Generate a map of places you have been/lived
}
