% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_images.R
\name{offset_images}
\alias{offset_images}
\title{Offset multiple Seurat images}
\usage{
offset_images(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\value{
A coordinate data frame
}
\description{
Internal function for defining coordinates with from multi-sample Seurat objects
}
\keyword{Bayesian}
\keyword{spatial}
\keyword{transcriptomics}
