% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_props_stacked.R
\name{plot_props_stacked}
\alias{plot_props_stacked}
\title{Plot stacked bar charts of mixture component proportions}
\usage{
plot_props_stacked(fit, group)
}
\arguments{
\item{fit}{A list returned by fit_maple()}

\item{group}{A column name of fit$W or a grouping vector of length N (nrow(fit$W))}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize the relative abundance of sub-populations after running fit_maple()
}
\keyword{Bayesian}
\keyword{spatial}
\keyword{transcriptomics}
