% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inla_dat.R
\name{prep_inla_dat}
\alias{prep_inla_dat}
\title{Format data to run INLA model.}
\usage{
prep_inla_dat(deaths, population, model, forecast.horizon)
}
\description{
Format data to run INLA model.
}
\keyword{internal}

