% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_row_summaries.R
\name{calculate_row_summaries}
\alias{calculate_row_summaries}
\title{Given a matrix, calculate row summaries}
\usage{
calculate_row_summaries(m)
}
\description{
Given a matrix, calculate row summaries
}

