% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_fit_ensemble.R
\name{maple_fit_ensemble}
\alias{maple_fit_ensemble}
\title{Fit an ensemble of models.}
\usage{
maple_fit_ensemble(deaths, population, forecast.horizon,
  models = maple_models(), num.draws = 1000, ax = NULL,
  num.threads = inla.getOption("num.threads"), verbose = TRUE)
}
\arguments{
\item{deaths}{Deaths matrix; see ?maple for more details.}

\item{population}{Population matrix; see ?maple for more details.}

\item{forecast.horizon}{The number of years to produce projections for.}

\item{models}{The individual models to be run; see ?maple_models for more details.}

\item{num.draws}{The number of posterior draws to sample and use for calculating statistical summaries.}

\item{ax}{The number of years lived on average by those who die in their current age group. See ?maple_plt for more details.}

\item{num.threads}{The number of threads passed to INLA; see ?maple for more details.}

\item{verbose}{If TRUE (the default), print some information on progress fitting models, etc.}
}
\value{
A list with the following entries
\describe{
  \item{sample.summaries}{A data frame holding statistical summary information for age-specific death rates, life expectancy and probability of dying, calculated from the posterior draws.}
  \item{samples}{A list of life table draws, calculated using posterior samples of death rates.}
  \item{model.fits}{The fitted model objects, which can be used for futher analysis.}
}
}
\description{
Fit an ensemble of models.
}
\examples{
fits <- maple_fit_ensemble(deaths = maple.deaths, population = maple.population,
                           forecast.horizon = 20, models = maple_models()[c(1, 20)])
}

