% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_plt.R
\name{maple_plt}
\alias{maple_plt}
\title{Generate a period life table for the supplied inputs. The ax for the open-ended age group 85+ is calculated using the Kannisto-Thatcher method [Thatcher et al, The survivor ratio method for estimating numbers at high ages, Demographic Research (6), 2002]. For age groups 5-9 to 80-84 the ax are calibrated using an iterative procedure described on p.47 of [Preston et al, Demography: measuring and modeling population processes, 2001].}
\usage{
maple_plt(death.rates, ax = NULL, check.conv = FALSE, full.table = FALSE)
}
\arguments{
\item{death.rates}{A matrix of death rates, with 18 rows, one for each 5-year age group 0-4, ..., 80-84, 85+.}

\item{ax}{The number of years lived on average by those who die in their current age group. For example, if all deaths in the age group 60-64 happened exactly at the middle of the age group, this would be equal to 2.5.}

\item{check.conv}{If TRUE, it will test that the iterative procedure to estimate ax (see above) has converged.}

\item{full.table}{If TRUE, returns all the columns of the period life table (see below).}
}
\value{
A period life table with death rates, life expectancies and probability of dying.
}
\description{
Generate a period life table for the supplied inputs. The ax for the open-ended age group 85+ is calculated using the Kannisto-Thatcher method [Thatcher et al, The survivor ratio method for estimating numbers at high ages, Demographic Research (6), 2002]. For age groups 5-9 to 80-84 the ax are calibrated using an iterative procedure described on p.47 of [Preston et al, Demography: measuring and modeling population processes, 2001].
}
\examples{
data(maple.deaths)
data(maple.population)
data(maple.ax)
plt <- maple_plt(maple.deaths / maple.population, maple.ax)
}

