% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_sample.R
\name{maple_sample}
\alias{maple_sample}
\title{Generate posterior samples of death rates and life tables. For models fitted with INLA, this uses INLA's inla.posterion.sample() function. For Lee Carter models, it samples from the distribution of the fitted parameters, using the estimated mean and standard deviation.}
\usage{
maple_sample(model.fit, num.draws, ax, ...)
}
\arguments{
\item{model.fit}{A model fitted using maple_fit_model().}

\item{num.draws}{The number of samples.}

\item{ax}{The number of years lived on average by those who die in their current age group. See ?maple_plt for more details.}
}
\value{
A list of life table draws, calculated using posterior samples of death rates.
}
\description{
Generate posterior samples of death rates and life tables. For models fitted with INLA, this uses INLA's inla.posterion.sample() function. For Lee Carter models, it samples from the distribution of the fitted parameters, using the estimated mean and standard deviation.
}

