% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maple_sample_summaries.R
\name{maple_sample_summaries}
\alias{maple_sample_summaries}
\title{Given a list of posterior samples of death rates and life tables, generate summaries for death rates, life expectancy at birth and at age 65, and probability of dying before age 70.}
\usage{
maple_sample_summaries(samples)
}
\arguments{
\item{samples}{Posterior sample draws produced with maple_sample().}
}
\value{
A data frame containing the mean, median, sd and 95% credible interval, for the following metrics
\describe{
  \item{rate_0, ..., rate_85:}{Age-specific death rates for age groups 0-4, ..., 80-84 and 85+.}
  \item{e0:}{Life expectancy at birth.}
  \item{e65:}{Life expectancy at age 65.}
  \item{q70:}{Probability of dying before age 70.}
}
}
\description{
Given a list of posterior samples of death rates and life tables, generate summaries for death rates, life expectancy at birth and at age 65, and probability of dying before age 70.
}

