% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_ex.R
\name{plt_ex}
\alias{plt_ex}
\title{Given a life table, extract the life expectancy values for a given age.}
\usage{
plt_ex(plt, x = 0)
}
\arguments{
\item{plt}{A life table calculated by maple_plt().}

\item{x}{The age group for which to extract life expectancy values.}
}
\value{
A named vector containing life expectancy values for each year.
}
\description{
Given a life table, extract the life expectancy values for a given age.
}

