% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_qx.R
\name{plt_qx}
\alias{plt_qx}
\title{Given a life table, extract the probability of dying by a given age.}
\usage{
plt_qx(plt, x = 70)
}
\arguments{
\item{plt}{A life table calculated by maple_plt().}

\item{x}{The age group for which to calculate probability of dying by.}
}
\value{
A named vector containing the probability of dying between 0 and age x in each year.
}
\description{
Given a life table, extract the probability of dying by a given age.
}

